+++ app/src/main/java/com/example/studyprojecttwo/domain/WeatherInfoShort.kt
@@ -1,5 +1,5 @@
 package com.example.studyprojecttwo.domain
 
 internal data class WeatherInfoShort(
-    val averageTemperature: Int,
+    val averageTemperature: Float,
 )
